+++ notes-app/src/Notes/NoteApp/addBtn.jsx
