+++ notes-app/src/Notes/NoteMain/Note.css
