+++ notes-app/src/Notes/SideBar/SideBar.jsx
